feat(neovim): improve auto-session and neo-tree integration
This commit addresses two issues related to the auto-session and
neo-tree plugins in Neovim.

Firstly, it fixes the issue where auto-session was saving a buffer named
a directory, causing Neovim to open an empty directory buffer every time
it was launched. This was resolved by configuring the sessionoptions
global and adding a function to suppress netrw, the built-in file
explorer in Neovim.

Secondly, it updates the neo-tree plugin to open the file browser when
no session is restored. This is achieved by creating an autocmd that
triggers on the 'AutoSession::NoSessionRestored' event.

The commit also includes updates to the lazy-lock.json file to reflect
the latest commits of the nvim-web-devicons and rustaceanvim plugins.